refactor(spaces): abstracted helpers from spaces and exported sanitize function
