+++ packages/spaces/index.d.ts
@@ -2,6 +2,7 @@ import Spaces, { useSpace } from './types/Spaces';
 import SpacesImage from './types/SpacesImage';
 import SpacesDisclaimer from './types/SpacesDisclaimer';
 import SpacesAgreement from './types/SpacesAgreement';
+import { sanitizeSpaces } from './types/helpers';
 
 export default Spaces;
 
@@ -12,6 +13,7 @@ export {
     SpacesImage as SpacesTile,
     SpacesDisclaimer,
     SpacesAgreement,
-    useSpace
+    useSpace,
+    sanitizeSpaces
 };
