+++ packages/spaces/types/helpers.d.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export function sanitizeSpaces(spaces: object[]): object[];
